Moved more basic operations to ops; Fixed bit and  mod operations to use the right toInt
